Moved entities and resource modules to a more descriptive directory
