+++ src/components/Admin/src/admin.vue
@@ -13,8 +13,8 @@
 <script>
 import Core from "@components/Core";
 import Ui from "@components/Ui";
-import resource from "@store/resource";
-import { entities } from "@store/entities"
+import resourceModule from "@store/modules/resource";
+import entitiesModule from "@store/modules/entities"
 import auth from '@store/auth'
 import user from '@store/user'
 import UI_CONTENT from '@constants/ui.content.default'
@@ -36,8 +36,8 @@ export default {
     Ui
   },
   created() {
-    this.$store.registerModule('resources', resource)
-    this.$store.registerModule('entities', entities)
+    this.$store.registerModule('resources', resourceModule)
+    this.$store.registerModule('entities', entitiesModule)
     this.$store.registerModule('auth', auth)
     this.$store.registerModule('user', user)
   },
